- updated to latest version - fixed compilation of cpu-implementations when cuda is'nt found on the system

+++ lib_ensembles/source/cpudte.h
@@ -35,7 +35,7 @@ class CpuDte : public lib_algorithms::MlAlgorithm<T> {
   void featureImportanceThread();
 
   T lnFunc(T num);
-  void Seed(int seed, bool bagging, int nr_samples, col_array<int>& indices,
+  void Seed(uint64_t seed, bool bagging, int nr_samples, col_array<int>& indices,
             int total_samples);
 };
 }
